refactor(core): split config parsing modules
